refactor(framework): share the run paths' driver-event accounting
A build and a direct prompt differ in what they do with the agent, but the
accounting around it is the same, and each file carried its own copy: the
session event, the session-id/usage tracking, the budget cap (#322) and the
consumption gate (#529). Stripped of comments the two onDriverEvent handlers
were 30 identical lines, differing only by a : void annotation.

This is the shape of #563, one site earlier: emitTurnSignals was byte-identical
in these same two files and the third turn site never got a copy. These copies
have not drifted yet, which is the point of extracting them now.

The teardown block is deliberately left alone: run.ts has a declined branch
(#358) and disposes a runner, so it differs for real.

Closes #565
